prior_auth/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
anthropic
python-dotenv
reportlab
